// SistemaVentasBatia/ClientApp/src/app/models/usuarioregistro.ts
export interface UsuarioRegistro {
    idAutorizacionVentas: number;
    idPersonal: number;
    autoriza: number;
    nombres: string;
    apellidos: string;
    puesto: string;
    telefono: string;
    telefonoExtension: string;
    telefonoMovil: string;
    email: string;
    firma: string;
    revisa: number;
}