// SistemaVentasBatia/ClientApp/src/app/models/immsjornada.ts
export interface ImmsJornada {
    idImmsJornadaCotizador: number;
    normal2: number;
    normal4: number;
    normal8: number;
    normal12: number;
    frontera2: number;
    frontera4: number;
    frontera8: number;
    frontera12: number;
    fechaAlta: Date;
    idPersonal: number;
    usuario: string;
}